refactored server.py have controllers group and graph refactored further with database and s3 connection code abstracted out and tested for working completely
