+++ Part2/Part2_Backend/database/__init__.py
@@ -1,0 +1,38 @@
+import os
+from sqlalchemy import create_engine
+from sqlalchemy.orm import sessionmaker
+from sqlalchemy.pool import NullPool
+from contextlib import contextmanager
+
+from dotenv import load_dotenv
+
+load_dotenv()
+
+if os.getenv("ENV") == "production":
+    engine = create_engine(
+        os.getenv("DATABASE_URL"),
+        poolclass=NullPool,
+        pool_pre_ping=True
+    )
+else:
+    engine = create_engine(
+        os.getenv("DATABASE_URL"),
+        pool_size=5,
+        max_overflow=10,
+        pool_recycle=300,
+        pool_pre_ping=True
+    )
+
+SessionLocal = sessionmaker(bind=engine)
+
+@contextmanager
+def session_scope():
+    session = SessionLocal()
+    try:
+        yield session
+        session.commit()
+    except Exception:
+        session.rollback()
+        raise
+    finally:
+        session.close()
