+++ Part2/Part2_Backend/database/crud.py
@@ -1,0 +1,36 @@
+from .models import Group, File
+
+def create_group(session, user_id, title, description, is_domain_specific, genomes, num_genes, num_domains):
+    group = Group(
+        user_id=user_id,
+        title=title,
+        description=description,
+        is_domain_specific=is_domain_specific,
+        genomes=genomes,
+        num_genes=int(num_genes),
+        num_domains=int(num_domains)
+    )
+    session.add(group)
+    session.flush() 
+    return group
+
+def add_file(session, group_id, user_id, file_name, s3_key, file_format):
+    file = File(
+        group_id=group_id,
+        user_id=user_id,
+        file_name=file_name,
+        s3_key=s3_key,
+        file_type=file_format
+    )
+    session.add(file)
+
+def get_first_or_none(session, model, **filters):
+    return session.query(model).filter_by(**filters).first()
+
+def get_all(session, model, **filters):
+    return session.query(model).filter_by(**filters).all()
+
+def delete(session, model_obj):
+    session.delete(model_obj)
+
+
